Rename to make clear these are PHPMailer integrations
